♻️ refactor: components에 shared폴더를 추가하여 공통 컴포넌트 구분
